+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.1 under development
 
 - Enh #17: Get request from provider into `RequestCookies` every time `get()` and `has()` methods are called (@vjik)
+- Chg #18: Add `RequestCookieProvider` instead of `RequestCookies`, which is marked as deprecated (@vjik)
 
 ## 1.1.0 October 28, 2024